Add Authorization struct
And add exposed DecodeAuthorization function.
And rename ValidPurchases to ActivePurchases.

In order to use the ActiveAuthorizationIDs notice from tunnel-core, authorizations must be decoded+parsed (so that IDs can be compared).

Additionally, clients with subscriptions need Authorization fields outside of the context of PsiCash. Instead of writing deserializers in every client, the PsiCash core library will expose a general-purpose decoder.