keep backup of datastore file, to help recover from corruption
We have seen data corruption in user feedback (usually in the form of JSON parsing failing, but sometimes in the form of impossible data values, we think). We can't be certain that this is due to file corruption, but it's certainly possible that out of millions of users, some will have file corruption issues each day. We believe (and hope) that having a backup datastore file brings that probability down greatly -- two files need to suffer the same improbable event at the same time.